have resource and datasource models together, renaming resource_schema files to schema
