+++ .golangci.yml
@@ -123,7 +123,7 @@ issues:
     - linters:
         - gocritic
       text: "^hugeParam: req is heavy"
-    - path: "_schema\\.go" # exclude rules for schema files as it's auto-genereated from OpenAPI spec
+    - path: "schema\\.go" # exclude rules for schema files as it's auto-genereated from OpenAPI spec
       text: "fieldalignment|hugeParam|var-naming|ST1003|S1007|exceeds the maximum|too long|regexpSimplify|nolint"
 run:
   timeout: 10m
